requirements-linux.txt
azure-cli==2.50.0
azure-identity==1.10.0
azure-kusto-data==4.2.0
colorama==0.4.6
cryptography==41.0.2
jinja2==3.1.4
msal==1.22.0
msal-extensions==1.0.0
packaging
pandas==2.0.3
pyodbc==4.0.39
pyjwt==2.8.0
python-dotenv==1.0.1
pytest==7.1.3
pyyaml
tabulate==0.9.0
urllib3==1.26.15
strenum==0.4.15
requests==2.31.0
tzlocal==5.2
numpy==1.24.4
